// src/typings/inputValueType.ts
export enum InputValueType {
  FLOAT = "float",
  INTEGER = "integer",
  LIST = "list",
  ARRAY = "array",
  DICT = "dict",
  STRING = "string",
  BOOLEAN = "boolean",
  JSON_SCHEMA = "json_schema",
  IMAGE = "image",
  BASE64_FILE = "base64_file",
  DOCLING_DOCUMENT = "docling_document",
  KNOWLEDGE_BASE = "knowledge_base",
  TOOLS = "tools",
  TOOL = "tool",
  WORKFLOW = "workflow",
  STRUCT = "struct"
}